Change the initial widths from percentage to pixels for proper column resizing
