+++ datahub-web-react/src/app/entity/shared/constants.ts
@@ -109,3 +109,7 @@ export enum EntityProfileTab {
     DOCUMENTATION_TAB = 'DOCUMENTATION_TAB',
     DATA_PRODUCTS_TAB = 'DATA_PRODUCTS_TAB',
 }
+
+export const SCHEMA_TABLE_MIN_COLUMN_WIDTH = 80;
+export const SCHEMA_TABLE_MAX_COLUMN_WIDTH = 800;
+export const SCHEMA_TABLE_MIN_COLUMN_HEIGHT = 30;
