+++ datahub-web-react/src/app/entity/shared/tabs/Dataset/Schema/SchemaTable.tsx
@@ -21,10 +21,11 @@ import { StyledTable } from '../../../components/styled/StyledTable';
 import { SchemaRow } from './components/SchemaRow';
 import { FkContext } from './utils/selectedFkContext';
 import useSchemaBlameRenderer from './utils/useSchemaBlameRenderer';
-import { ANTD_GRAY } from '../../../constants';
+import { ANTD_GRAY, SCHEMA_TABLE_MAX_COLUMN_WIDTH } from '../../../constants';
 import MenuColumn from './components/MenuColumn';
 import translateFieldPath from '../../../../dataset/profile/schema/utils/translateFieldPath';
 import { ResizableTitle } from './ResizableTitle';
+import { percentToPixelWidth } from '../../../utils';
 
 const TableContainer = styled.div`
     overflow: inherit;
@@ -58,7 +59,6 @@ export type Props = {
 
 const EMPTY_SET: Set<string> = new Set();
 const TABLE_HEADER_HEIGHT = 52;
-const MAX_COLUMN_WIDTH = 600;
 
 export default function SchemaTable({
     rows,
@@ -116,7 +116,7 @@ export default function SchemaTable({
     });
 
     const fieldColumn = {
-        width: '22%',
+        width: percentToPixelWidth('22%'),
         title: 'Field',
         dataIndex: 'fieldPath',
         key: 'fieldPath',
@@ -128,7 +128,7 @@ export default function SchemaTable({
     };
 
     const descriptionColumn = {
-        width: '22%',
+        width: percentToPixelWidth('22%'),
         title: 'Description',
         dataIndex: 'description',
         key: 'description',
@@ -136,7 +136,7 @@ export default function SchemaTable({
     };
 
     const tagColumn = {
-        width: '13%',
+        width: percentToPixelWidth('13%'),
         title: 'Tags',
         dataIndex: 'globalTags',
         key: 'tag',
@@ -145,7 +145,7 @@ export default function SchemaTable({
     };
 
     const termColumn = {
-        width: '13%',
+        width: percentToPixelWidth('13%'),
         title: 'Glossary Terms',
         dataIndex: 'globalTags',
         key: 'tag',
@@ -154,7 +154,7 @@ export default function SchemaTable({
     };
 
     const blameColumn = {
-        width: '10%',
+        width: percentToPixelWidth('10%'),
         dataIndex: 'fieldPath',
         key: 'fieldPath',
         render(record: SchemaField) {
@@ -178,7 +178,7 @@ export default function SchemaTable({
     }
 
     const usageColumn = {
-        width: '10%',
+        width: percentToPixelWidth('10%'),
         title: 'Usage',
         dataIndex: 'fieldPath',
         key: 'usage',
@@ -187,7 +187,7 @@ export default function SchemaTable({
     };
 
     const menuColumn = {
-        width: '5%',
+        width: percentToPixelWidth('5%'),
         title: '',
         dataIndex: '',
         key: 'menu',
@@ -228,7 +228,7 @@ export default function SchemaTable({
         (e, { size }) => {
             setColumns((prevColumns) => {
                 const nextColumns = [...prevColumns];
-                const newWidth = Math.min(size.width, MAX_COLUMN_WIDTH);
+                const newWidth = Math.min(size?.width, SCHEMA_TABLE_MAX_COLUMN_WIDTH);
                 nextColumns[index] = {
                     ...nextColumns[index],
                     width: newWidth,
